continuation mini jeu, on est sur la bonne voie

+++ Assets/Scripts/Difficulty/DifficultyData.cs
@@ -14,7 +14,7 @@ namespace Enemy
         public float PlayerStrengthMultiplier { get; private set; }
         
         [field : SerializeField]
-        public float PlayerHealthBarAmountMultiplier { get; private set; }
+        public int PlayerHealthBarAmountMultiplier { get; private set; }
         #endregion
         
         #region Enemy
